requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
